services/alert/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
asyncpg==0.29.0
pydantic==2.7.1
prometheus-client==0.20.0
python-dotenv==1.0.1
